Fixes for linux64 compiler warnings following Win64 port.

+++ src/SecondOrderAssocReaction.cpp
@@ -155,8 +155,8 @@ namespace mesmer
 
     // Calculate the elements of the product block.
 
-    const int pdtLocation = isomermap[m_pdt1];
-    const int pdtBasisSize = static_cast<int>(m_pdt1->getColl().get_nbasis());
+    const size_t pdtLocation = isomermap[m_pdt1];
+    const size_t pdtBasisSize = m_pdt1->getColl().get_nbasis();
     for (size_t i = 0, ii(pdtLocation), egvI(pdtColloptrsize - 1); i < pdtBasisSize; i++, ii++, --egvI) {
       (*CollOptr)[ii][ii] -= m_pdt1->getColl().matrixElement(egvI, egvI, RvsMicroRateCoef);
       for (size_t j = i + 1, jj(pdtLocation + j), egvJ(pdtColloptrsize - j - 1); j < pdtBasisSize; j++, jj++, --egvJ) {
